+++ flake8_django/checker.py
@@ -1,8 +1,9 @@
 import ast
 
-from flake8_django.checkers import ModelDunderStrMissingChecker, ModelFieldChecker, ModelFormChecker, URLChecker, RenderChecker
+from flake8_django.checkers import ModelDunderStrMissingChecker, ModelFieldChecker, ModelFormChecker, \
+    URLChecker, RenderChecker, ModelMetaChecker
 
-__version__ = '0.0.3'
+__version__ = '0.0.5'
 
 
 class DjangoStyleFinder(ast.NodeVisitor):
@@ -18,6 +19,7 @@ class DjangoStyleFinder(ast.NodeVisitor):
         'ClassDef': [
             ModelFormChecker(),
             ModelDunderStrMissingChecker(),
+            ModelMetaChecker(),
         ]
     }